Extending Sleeper to be configurable
A nice feature would be for the `Sleeper` to be configurable.
This means that we can adjust the sleep time in our main program.
This can achieved by creating a new type: `ConfigurableSleeper` that
accepts what we need for configuration and testing.

We are using `duration` to configure the timeslept and `sleep` as a way
to pass in a sleep function. The signature of `sleep` is the same as for
`time.Sleep` allowing us to use `time.Sleep` in our real implementaion
and the following custom spy in our tests. With this custom spy in
place, we can also create a new test for the configurable sleeper.