non-parallel website checker
The benchmark testing using a slice of one hundred urls and a fake
implementation WebsiteChecker which is deliberately show, the one which
uses time.Sleep to wait exactly twenty milliseconds and then it returns
true.

When we run the benchmark using go test -bench=.
CheckWebsites has been benchmarked at 2e9 ns - about two and a quarter
seconds.

In the next iteration, let's try and make it faster