using goroutines and channels
An operation that does not block in Go will run in a seperate process
called a goroutine. For solving the data race (/race condition:
concurrent writes to map), we can co-ordincate our goroutines using
channels. Channels are a Go data structure that can both receive and
send values. These operations, along with their detuals, allow
communication between different processes.

In this case we want to think about the communication between the parent
process and each of the goroutines that it makes to do the work of
running WebsiteChecker function with the url.

Instead of writing to the map directly we're sending a result struct for
each call to wc to the resultChannel with a send statement. This uses
the `<-` operator, taking a channel on the left and a value on the
right.

// Send statement
resultChannel <- result{url_string: u, status: wc(u)}

// Receive statement
r := <- resultChannel

^see how the two operands are reversed in case of receive statement. The
channel is now on the right and the variable that we're assigning to is
on the left.

When we run the benchmark now, we get 2e7 ns - 0.023 seconds, about one
hundred times as fast as the original function. A great success!